File-based CDK: reorganize FileReadMode to fix circular import
